requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
aiofiles>=23.2.1
